# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RetinaGuard AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type PatientGender = (typeof PatientGender)[keyof typeof PatientGender];

export const PatientGender = {
  male: "male",
  female: "female",
  other: "other",
} as const;

/**
 * @nullable
 */
export type PatientDiabetesType =
  | (typeof PatientDiabetesType)[keyof typeof PatientDiabetesType]
  | null;

export const PatientDiabetesType = {
  type1: "type1",
  type2: "type2",
  gestational: "gestational",
} as const;

export interface Patient {
  id: number;
  name: string;
  age: number;
  gender: PatientGender;
  /** @nullable */
  diabetesType: PatientDiabetesType;
  /** @nullable */
  contactInfo: string | null;
  createdAt: string;
}

export type CreatePatientInputGender =
  (typeof CreatePatientInputGender)[keyof typeof CreatePatientInputGender];

export const CreatePatientInputGender = {
  male: "male",
  female: "female",
  other: "other",
} as const;

/**
 * @nullable
 */
export type CreatePatientInputDiabetesType =
  | (typeof CreatePatientInputDiabetesType)[keyof typeof CreatePatientInputDiabetesType]
  | null;

export const CreatePatientInputDiabetesType = {
  type1: "type1",
  type2: "type2",
  gestational: "gestational",
} as const;

export interface CreatePatientInput {
  name: string;
  /**
   * @minimum 1
   * @maximum 120
   */
  age: number;
  gender: CreatePatientInputGender;
  /** @nullable */
  diabetesType?: CreatePatientInputDiabetesType;
  /** @nullable */
  contactInfo?: string | null;
}

export type ScanRiskLevel = (typeof ScanRiskLevel)[keyof typeof ScanRiskLevel];

export const ScanRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Scan {
  id: number;
  patientId: number;
  /** @nullable */
  patientName: string | null;
  /** @nullable */
  imageData: string | null;
  /**
   * @minimum 0
   * @maximum 4
   */
  drStage: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  confidenceScore: number;
  riskLevel: ScanRiskLevel;
  /**
   * @minimum 0
   * @maximum 100
   */
  blindnessRiskScore: number;
  /** @nullable */
  heatmapData: string | null;
  doctorConfirmed: boolean;
  /** @nullable */
  doctorNotes: string | null;
  /** @nullable */
  doctorId: string | null;
  recommendation: string;
  createdAt: string;
}

export interface AnalyzeScanInput {
  patientId: number;
  /** Base64-encoded image or image URL */
  imageData: string;
}

export interface UpdateScanInput {
  /** @nullable */
  doctorConfirmed?: boolean | null;
  /** @nullable */
  doctorNotes?: string | null;
  /** @nullable */
  doctorId?: string | null;
}

export interface Report {
  scan: Scan;
  patient: Patient;
  generatedAt: string;
  drStageName: string;
  riskDescription: string;
  treatmentPlan: string;
}

export interface DrStageDistribution {
  stage: number;
  stageName: string;
  count: number;
}

export interface RiskBreakdown {
  level: string;
  count: number;
  percentage: number;
}

export interface AnalyticsSummary {
  totalPatients: number;
  totalScans: number;
  confirmedScans: number;
  criticalCases: number;
  averageConfidence: number;
  drStageDistribution: DrStageDistribution[];
  riskBreakdown: RiskBreakdown[];
  recentScans: Scan[];
}

export type DeletePatient200 = {
  success?: boolean;
  message?: string;
};

export type GetScansParams = {
  /**
   * @nullable
   */
  patientId?: number | null;
};

export type DeleteScan200 = {
  success?: boolean;
  message?: string;
};